[PATCH] Implement asynchronous suspend/resume methods for system  module
Especially in Android based systems suspend isn't synchronous. They are working with
wakelocks and therefor the request to suspend is only an intent and the device will not
suspend as long as no wakelock is hold anymore. The system needs to detect manually
when it should switch back to a alive state. With the new API the power management
implementation can request suspending the system with calling the new suspend_async
method and resuming through the resume method.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Herman van Hazendonk <github.com@herrie.org>